Fix issue that "open here" mapping sets virtualedit=all.
<Leader>oh is a mapping for opening a new line, starting at the same
column the cursor is at. I used virtualedit to accomplish that, but in
the old, messier version of this mapping, left virtualedit set to all
instead of turning it off afterward. Note that this is still a bit of a
messy thing because if virtualedit was on beforehand, this mapping turns
it off.